backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1  # MongoDB driver

# Authentication and Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.6  # Form data parsing

# Data validation
pydantic==2.5.0
pydantic[email]==2.5.0  # Email validation

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # Testing HTTP client

# Environment variables
python-dotenv==1.0.0
